requirements.txt
farm-haystack
farm-haystack[elasticsearch]
streamlit